Module Ocannl

module Nn_blocks : sig ... end

Prior to OCANNL 0.5, this module is just a placeholder hinting at an intended design pattern for model components.

module Operation : sig ... end

Computational primitives for neural networks, integrating Tensor with Assignments.

module PrintBox_utils : sig ... end

module Row : sig ... end

The row type, shape inference related types and constraint solving.

module Shape : sig ... end

module Tensor : sig ... end

module Train : sig ... end