Module Operation.At

val (.@{}) : Tensor.t -> Base.int Base.array -> Base.float

Get the value at the given indices.

val (.@%{}) : Tensor.t -> Base.int Base.array -> Base.float

val (.@{}<-) : Tensor.t -> Base.int Base.array -> Base.float -> Base.unit

Set the value at the given indices.

val (.@%{}<-) : Tensor.t -> Base.int Base.array -> Base.float -> Base.unit

val (.@[]) : Tensor.t -> Base.int -> Base.float

Get the value at the given index from a single-axis shape tensor.

val (.@%[]) : Tensor.t -> Base.int -> Base.float

val (.@[]<-) : Tensor.t -> Base.int -> Base.float -> Base.unit

Set the value at the given index for a single-axis shape tensor.

val (.@%[]<-) : Tensor.t -> Base.int -> Base.float -> Base.unit