Module Operation.TDSL

include module type of struct include Initial_TDSL end

val term : label:Base.string Base.list -> ?batch_dims:Base.int Base.list -> ?input_dims:Base.int Base.list -> ?output_dims:Base.int Base.list -> ?batch_axes:(Base.string * Base.int) Base.list -> ?input_axes:(Base.string * Base.int) Base.list -> ?output_axes:(Base.string * Base.int) Base.list -> ?deduced:Shape.deduce_within_shape -> ?init_op:Tensor.init_op -> ?fetch_op:(v:Tensor.tn -> Tensor.fetch_op) -> Base.unit -> Tensor.t

val number : ?label:Base.string Base.list -> ?axis_label:Base.string -> Base.float -> Tensor.t

val ndarray : ?label:Base.string Base.list -> ?batch_dims:Base.int Base.list -> ?input_dims:Base.int Base.list -> ?output_dims:Base.int Base.list -> ?batch_axes:(Base.string * Base.int) Base.list -> ?input_axes:(Base.string * Base.int) Base.list -> ?output_axes:(Base.string * Base.int) Base.list -> ?strict:Base.bool -> Base.float Base.array -> Tensor.t

val param : ?input_dims:Base.int Base.list -> ?output_dims:Base.int Base.list -> ?input_axes:(Base.string * Base.int) Base.list -> ?output_axes:(Base.string * Base.int) Base.list -> ?deduced:Shape.deduce_within_shape -> ?strict:Base.bool -> ?values:Base.float Base.array -> Base.string -> Tensor.t

module O = DO

val einsum : ?label:Base.string list -> Base.string -> Tensor.t -> Tensor.t -> Tensor.t

val outer_sum : ?label:Base.string list -> Base.string -> Tensor.t -> Tensor.t -> Tensor.t

val einsum1 : ?label:Base.string list -> Base.string -> Tensor.t -> Tensor.t

val range : ?label:Base.string list -> ?axis_label:Base.string -> Base.Int.t -> Tensor.t

val range_of_shape : ?label:Base.string list -> ?batch_dims:Base.Int.t Base.List.t -> ?input_dims:Base.Int.t Base.List.t -> ?output_dims:Base.Int.t Base.List.t -> ?batch_axes:(Base.string * Base.Int.t) Base.List.t -> ?input_axes:(Base.string * Base.Int.t) Base.List.t -> ?output_axes:(Base.string * Base.Int.t) Base.List.t -> unit -> Tensor.t

val stop_gradient : ?label:Base.string list -> Tensor.t -> Tensor.t

val init_const : l:Base.string -> ?strict:Base.bool -> ?b:Base.int Base.list -> ?i:Base.int Base.list -> o:Base.int Base.list -> Base.float Base.array -> Tensor.t

The input i dimensions default to empty. The batch dimensions will be inferred if omitted. strict controls whether Constant_fill will try to fit the given values in the tensor and contribute to shape inference. If it is not provided explicitly, it will be true if b is omitted, and false otherwise.

val init_param : l:Base.string -> ?b:Base.int Base.list -> ?i:Base.int Base.list -> ?o:Base.int Base.list -> Base.float Base.array -> Tensor.t

It's like `Tensor.param` but without shape inference.