Module Ocannl.Nn_blocks

This file contains basic building blocks for neural networks, with limited functionality. Feel free to copy-paste and modify as needed.

We follow "the principle of least commitment": where possible, we use row variables to remain agnostic to the number of axes. This flexibility often remains unused, but it makes explicit the architectural structure.

module Tn = Ir.Tnode

val mlp_layer : label:Base.string Base.list -> hid_dim:Base.int -> unit -> Tensor.t -> Tensor.t

val dropout : rate:Base.Float.t -> unit -> Tensor.t -> Tensor.t

Set rate=0.0 during inference.

val mlp : label:Base.string Base.list -> hid_dims:Base.int Base.List.t -> unit -> Tensor.t -> Tensor.t

Multi-layer perceptron of depth List.length hid_dims + 1, with a linear output layer.

val reduce_specified_axes : Base.String.t -> Base.String.t

val softmax : spec:Base.String.t -> ?temperature:Base.Float.t -> unit -> Tensor.t -> Tensor.t

Softmax across specified axes. Does not support non-default row variables.

val multi_head_attention : label:Base.string Base.list -> num_heads:Base.int -> ?temperature:Base.Float.t -> ?dropout_rate:Base.Float.t -> unit -> ?mask:Tensor.t -> Tensor.t -> Tensor.t

val layer_norm : label:Base.string Base.list -> ?epsilon:Base.float -> unit -> Tensor.t -> Tensor.t

val transformer_encoder_block : label:Base.string Base__List.t -> num_heads:Base.int -> d_ff:Base.int -> ?epsilon:Base.float -> unit -> Tensor.t -> Tensor.t

val cross_attention : label:Base.string Base.list -> num_heads:Base.int -> ?temperature:Base.Float.t -> ?dropout_rate:Base.Float.t -> unit -> Tensor.t -> enc_output:Tensor.t -> Tensor.t

val transformer_decoder_block : label:Base.string Base__List.t -> num_heads:Base.int -> d_ff:Base.int -> ?epsilon:Base.float -> unit -> Tensor.t -> enc_output:Tensor.t -> mask:Tensor.t -> Tensor.t

val transformer_encoder : label:Base.String.t Base__List.t -> num_layers:int -> num_heads:Base.int -> d_ff:Base.int -> ?epsilon:Base.float -> unit -> Tensor.t -> Tensor.t

val transformer_decoder : label:Base.String.t Base__List.t -> num_layers:int -> num_heads:Base.int -> d_ff:Base.int -> ?epsilon:Base.float -> unit -> Tensor.t -> enc_output:Tensor.t -> mask:Tensor.t -> Tensor.t

val transformer : label:Base.string Base.list -> num_encoder_layers:int -> num_decoder_layers:int -> num_heads:Base.int -> d_model:Base.int -> d_ff:Base.int -> ?epsilon:Base.float -> unit -> Tensor.t -> Tensor.t -> Tensor.t -> Tensor.t