Module Ocannl.Train

module Tn = Arrayjit.Tnode

module Nd = Arrayjit.Ndarray

module NTDSL = Operation.NTDSL

module Asgns = Arrayjit.Assignments

module Idx = Arrayjit.Indexing

module Utils = Arrayjit.Utils

module Rand = Arrayjit.Rand.Lib

module type Backend_type = Arrayjit.Backends.Backend

module Debug_runtime = Arrayjit.Utils.Debug_runtime

module CDSL : sig ... end

module IDX : sig ... end

val debug_rt : (module Minidebug_runtime.Debug_runtime)

val run : 'a Arrayjit.Backend_types.routine -> Base.unit

val fresh_backend : ?backend_name:Base.String.t -> ?config:Arrayjit.Backend_types.config -> unit -> (module Arrayjit__Backends.Backend)

Reinitializes a backend selected via a global backend flag.

val is_param : Tensor.t -> Base.bool

val get_params : Tensor.t -> (Tensor.t, Tensor.comparator_witness) Base.Set.t

val save_params : Tensor.t -> unit

val restore_params : Tensor.t -> unit

val set_on_host : Tn.memory_type -> Tn.t -> unit

val set_materialized : Tn.t -> unit

val set_hosted : Tn.t -> unit

val label_suffix : Base.String.t Base.List.t -> Base.String.t

val forward : ?disable_rootness_check:bool -> Tensor.t -> Asgns.t

Sets the tensor's value as "fully on host", returns the tensor's forward code with a label-derived comment.

type updaten = {

loss : Tensor.t;

label : Base.string;

params : (Tensor.t, Tensor.comparator_witness) Base.Set.t;

fwd_bprop : Asgns.t;

}

val grad_update : ?disable_rootness_check:bool -> ?setup_for_parallel:bool -> Tensor.t -> updaten

Returns the tensor's forward, zeroing gradients, and backprop code wrapped with label-derived comments. Sets the tensor's value as "fully on host". If setup_for_parallel is true (false by default), sets the parameters and their gradients as "non-local" (on-device).

val sgd_one : learning_rate:Tensor.t -> ?momentum:Base.Float.t -> ?weight_decay:Base.float -> ?nesterov:bool -> Tensor.t -> Asgns.t

See: https://github.com/tinygrad/tinygrad/blob/master/tinygrad/nn/optim.py

val sgd_update : learning_rate:Tensor.t -> ?momentum:Base.Float.t -> ?weight_decay:Base.float -> ?nesterov:bool -> updaten -> Asgns.t

val sequential_loop : f:(unit -> Base.unit) -> (Idx.static_symbol * int Base.ref) list -> Base.unit

All and only bindings with associated ranges are iterated, with the binding's initial value lost. Bindings without ranges remain at their initial values.

val round_robin : (unit -> unit) Base.Array.t -> Idx.lowered_bindings Base.Array.t -> (Idx.static_symbol * Base.int Base.ref) list -> sync:(Base.int -> Base.unit) -> Base.unit

Distributes iterated indices to workers in a round-robin fashion. All and only bindings with associated ranges are iterated, with the binding's initial value lost. Bindings without ranges remain at their initial values. sync is called after each round of calling all workers, and at the end if needed, with the number of workers called during the round.

val round_robin_dry_run : num_devices:Base__Int.t -> (Idx.static_symbol * int Base.ref) list -> dry_sync:(Base__Int.t -> Base.unit) -> Base.unit

val set_virtual : Tn.t -> unit

val every_non_literal_on_host : Tensor.t -> Base.unit

val all_host_to_device : (module Backend_type with type context = 'context) -> 'context0 -> Tensor.t -> Base.unit

val all_device_to_host : (module Backend_type with type context = 'context) -> 'context0 -> Tensor.t -> Base.unit

val sync_run : ?looping:(unit -> Base.unit) -> (module Backend_type with type context = 'context) -> 'context0 Arrayjit.Backend_types.routine -> Tensor.t -> Base.unit

Executes the jitted code and copies arrays embedded in the given tenosor from and to host, synchronizes before copying to host. If looping is provided, loops over bindings and executes the given function inside the loop after a run. All and only bindings with associated ranges are iterated, with the binding's initial value lost. Bindings without ranges remain at their initial values.

module Lazy = Utils.Lazy

let merge ?name_prefix tn ~accum ~(src : context) : code option =
  let bindings = Indexing.Empty in
  merge ?name_prefix tn ~accum ~src bindings |> Option.map ~f:(fun routine : code -> Compiled routine)

let merge_batch ?name_prefixes ~occupancy tns ~accum ~(srcs : context array) =
  let bindings = Indexing.Empty in
  merge_batch ?name_prefixes ~occupancy tns ~accum ~srcs bindings
  |> Hashtbl.map ~f:(fun procs ->
         Array.map procs ~f:(Option.map ~f:(fun routine : code -> Compiled routine)))

let%track_sexp merge_from_global ~unoptim_ll_source ~ll_source ~name ~dst ~accum ~src =
  let global =
    match src with
    | None -> Arrayjit.Ops.Merge_buffer src.id
    | Some src -> External_unsafe { ptr = src; prec = dst.Arrayjit.Tnode.prec; dims = dst.dims }
  in
  let open Arrayjit.Low_level in
  let body idcs =
    Set
      {
        tn = dst;
        idcs;
        llv = Binop (accum, Get (dst, idcs), Get_global (global, Some idcs));
        debug = "";
      }
  in
  let llc = loop_over_dims (Lazy.force dst.dims) ~body in
  optimize_proc ~unoptim_ll_source ~ll_source ~name [] llc

let compile_merges ?name_prefixes:_ ~occupancy:_
    (* Tnode.t -> dst_n:int -> dst:context -> src_n:int -> src:context -> Utils.requirement *) _tns
    ~accum:_ ~dsts:_ ~srcs:_ =
  failwith "NOT IMPLEMENTED YET"

let%track_sexp merge (type context) ?name_prefix la ~accum ~(src : context) bindings =
  (* FIXME: reconstruct name if missing *)
  let name = Option.value name_prefix ~default:"" in
  let unoptim_ll_source = Utils.get_debug_formatter ~fname:(name ^ "-unoptimized.ll") in
  let ll_source = Utils.get_debug_formatter ~fname:(name ^ ".ll") in
  let name_prefix : string = Option.value ~default:"" @@ Option.map name_prefix ~f:(fun s -> s ^ "_") in
  Option.map (Map.find src.arrays la) ~f:(fun (src : Ndarray.t) ->
      let name = [%string "%{name_prefix}into_%{Tn.name la}"] in
      compile ~opt_ctx_arrays:None ~name bindings
      @@ merge_from_global ~unoptim_ll_source ~ll_source ~name ~dst:la ~accum
           ~src:(Ndarray.get_voidptr src))

let%track_sexp merge_batch ?(name_prefixes : string array option) ~occupancy tns ~accum
    ~(srcs : context array) bindings =
  (* FIXME: reconstruct name if missing *)
  let name =
    String.(
      strip ~drop:(equal_char '_')
      @@ common_prefix @@ Array.to_list
      @@ Option.value name_prefixes ~default:[||])
  in
  let unoptim_ll_source = Utils.get_debug_formatter ~fname:(name ^ "-unoptimized.ll") in
  let ll_source = Utils.get_debug_formatter ~fname:(name ^ ".ll") in
  let complete =
    Array.concat_map (Array.of_list tns) ~f:(fun tn ->
        Array.mapi srcs ~f:(fun i src ->
            match (occupancy tn ~src_n:i ~src, Map.find src.arrays tn) with
            | Utils.Skip, _ -> ((tn, i), (None, None))
            | Optional { callback_if_missing }, None ->
                callback_if_missing ();
                ((tn, i), (None, None))
            | Required, None ->
                failwith @@ "Gccjit_backend.merge_batch: missing tnode " ^ Tn.name tn ^ " in context "
                ^ src.label
            | _, Some src ->
                let prefix = match name_prefixes with Some ns -> ns.(i) ^ "_" | None -> "" in
                let name = [%string "%{prefix}into_%{Tn.name tn}"] in
                ( (tn, i),
                  ( Some name,
                    Some
                      (merge_from_global ~unoptim_ll_source ~ll_source ~name ~dst:tn ~accum
                         ~src:(Ndarray.get_voidptr src)) ) )))
  in
  let ids, compileds = Array.unzip complete in
  let len = Array.length srcs in
  let together =
    Hashtbl.of_alist_exn (module Tn) @@ List.map tns ~f:(fun tn -> (tn, Array.create ~len None))
  in
  if Array.is_empty compileds then together
  else
    let names, compileds = Array.unzip compileds in
    let result = compile_batch ~opt_ctx_arrays:None ~names bindings compileds in
    Array.iter2_exn ids result ~f:(fun (tn, i) res ->
        let r = Hashtbl.find_exn together tn in
        assert (Option.is_none r.(i));
        r.(i) <- res);
    together

val collapse_merges : ('a, 'b Base.Option.t Base.Array.t) Base.Hashtbl.t -> 'c Base__List.t Base.Array.t

val parallel_update : (module Backend_type with type context = 'context) -> grad_updates:'context0 Arrayjit.Backend_types.routine Base.array -> sgd_update:'context1 Arrayjit.Backend_types.routine -> post_sync:(num_synced_devices:Base.int -> Base.unit) -> updaten -> Base.unit -> Base.unit

Performs one optimization step, potentially in parallel (if grad_updates are compiled for different devices). All jitted code must have the same bindings. Iterates over bindings with ranges, calling one of grad_updates in a round-robin fashion, and performs the following synchronization each time all grad_updates have been called:

1. merges all gradients into the device of grad_updates.(0), 2. calls sgd_update, 3. copies all parameters from the grad_updates.(0) device to the other devices, if needed, 4. calls post_sync with the number of devices synced since the previous sync.

All and only bindings with associated ranges are iterated, with the binding's initial value lost. Bindings without ranges remain at their initial values.

val debug_name : Tensor.t -> Base.String.t

val get_all_suggested_devices : ?max_num_devices:Base.Int.t -> (module Backend_type with type device = 'device) -> 'device0 Base.array

val example_train_loop : ?disable_rootness_check:bool -> seed:Base.int -> batch_size:Base__Int.t -> init_lr:Base.float -> ?lr_schedule: (batch_n:Idx.static_symbol -> step_n:Idx.static_symbol -> Tensor.t) -> ?max_num_devices:Base.Int.t -> data_len:Base__Int.t -> epochs:Base__Int.t -> inputs:(b:Base__Int.t list -> Tensor.t) -> outputs:(b:Base__Int.t list -> Tensor.t) -> model:(Tensor.t -> Tensor.t) -> loss_fn:(output:Tensor.t -> expectation:Tensor.t -> Tensor.t) -> weight_decay:Base.float -> ?per_batch_callback: (at_batch:Base.int -> at_step:Base.int -> learning_rate:Base.float -> batch_loss:Base.float -> epoch_loss:float -> unit) -> ?per_epoch_callback: (at_step:Base.int -> at_epoch:int -> learning_rate:Base.float -> epoch_loss:float -> unit) -> (module Backend_type) -> unit -> Tensor.t * Tensor.t * Tensor.t * (Base.float Base.array -> Base.float Base.array) * Base.float list * float list * Base.float list

val forward_and_forget : ?disable_rootness_check:Base.bool -> (module Backend_type with type context = 'context) -> 'context0 -> ?bindings:(Base.unit -> Base.unit) Idx.bindings -> Tensor.t -> Base.unit